refactor class names and removed the license duplication
