+++ twister2/examples/src/java/edu/iu/dsc/tws/examples/internal/task/batch/ReduceBatchTask.java
@@ -50,8 +50,8 @@ import edu.iu.dsc.tws.rsched.core.SchedulerContext;
 import edu.iu.dsc.tws.task.api.IFunction;
 import edu.iu.dsc.tws.task.api.IMessage;
 import edu.iu.dsc.tws.task.api.TaskContext;
-import edu.iu.dsc.tws.task.batch.AbstractBatchSinkTask;
-import edu.iu.dsc.tws.task.batch.AbstractBatchSourceTask;
+import edu.iu.dsc.tws.task.batch.BaseBatchSinkTask;
+import edu.iu.dsc.tws.task.batch.BaseBatchSourceTask;
 import edu.iu.dsc.tws.task.graph.DataFlowTaskGraph;
 import edu.iu.dsc.tws.task.graph.GraphBuilder;
 import edu.iu.dsc.tws.task.graph.OperationMode;
@@ -95,7 +95,7 @@ public class ReduceBatchTask implements IWorker {
     executor.execute();
   }
 
-  private static class GeneratorTask extends AbstractBatchSourceTask {
+  private static class GeneratorTask extends BaseBatchSourceTask {
     private static final long serialVersionUID = -254264903510284748L;
     private TaskContext sourceTaskContext;
     private Config config;
@@ -126,7 +126,7 @@ public class ReduceBatchTask implements IWorker {
     }
   }
 
-  private static class RecevingTask extends AbstractBatchSinkTask {
+  private static class RecevingTask extends BaseBatchSinkTask {
     private static final long serialVersionUID = -254264903510284798L;
     private int count = 0;
     private Config config;
