+++ twister2/task/src/main/java/edu/iu/dsc/tws/task/streaming/BaseStreamSourceTask.java
@@ -10,32 +10,21 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
 package edu.iu.dsc.tws.task.streaming;
 
 import edu.iu.dsc.tws.common.config.Config;
-import edu.iu.dsc.tws.task.api.IMessage;
-import edu.iu.dsc.tws.task.api.ISink;
+import edu.iu.dsc.tws.task.api.ISource;
 import edu.iu.dsc.tws.task.api.TaskContext;
 
-public abstract class AbstractStreamSinkTask implements ISink {
+public abstract class BaseStreamSourceTask implements ISource {
+
   private static final long serialVersionUID = -254264120110286748L;
   protected TaskContext ctx;
   protected Config config;
 
   @Override
-  public boolean execute(IMessage message) {
-    return false;
+  public void run() {
+
   }
 
   @Override
